Rock, Paper, Scissors Best Out of Five
This is a modification of the rock, paper, scissors code developed by Al Sweigart in his book "Automate the Boring Stuff with Python, 2nd Edition." The vast majority of my modifications appear from line 76 onward.

The modifications will end the game when either the user or the computer win three rounds. From there, the user will be asked if they would like to play again. Responding with yes will reset the win/loss tally and restart the game.